illustrate tent by image corruption example
to illustrate the tent method and fully test-time adaptation setting,
we provide an example for adaptation to image corruptions.

this is simply *example code* for explanation, not *reference code* for
reproduction. that said, experimenting with this should give results
that are representative.

reference code will follow to reproduce our ImageNet-C results